Session and logout button handled by nav.js
