+++ src/index.html
@@ -23,6 +23,7 @@
 <body>
   <header>
     <nav class="navbar navbar-expand-lg bg-light navbar-light fixed-top">
+      <script type="module" src="./js/nav.js"></script>
       <div class="container">
         <a class="navbar-brand text-uppercase fw-bold" href="index.html">
           <img src="img/wikipstudy_small-logo.png" height="50" width="50" alt="WikipStudy logo"
