+++ src/js/index.js
@@ -1,29 +1,9 @@
 import { quizzesByOwner } from "../graphql/queries.js";
-import {
-  addQuizCard,
-  appendAlert,
-  deleteCookie,
-  getCookie,
-  sessionCheck,
-} from "./utils.js";
+import { addQuizCard, appendAlert, getCookie } from "./utils.js";
 import graphqlCall from "../graphql/graphqlCall.js";
 
-sessionCheck().then((user) => {
-  if (!user) {
-    window.location.href = "login.html";
-  }
-});
-
 const responseMessage = document.getElementById("response-message");
-const logoutButton = document.getElementById("logout");
 const quizzesListElement = document.getElementById("quizzes-list");
-
-logoutButton.addEventListener("click", async () => {
-  deleteCookie("token");
-  deleteCookie("user_name");
-  deleteCookie("id");
-  window.location.href = "login.html";
-});
 
 const response = await graphqlCall(quizzesByOwner, { owner: getCookie("id") });
 if (!response.ok) {
